add support for REFSYM, and its weird "hidden data" which used to break parsing this "hidden data" (so is named in the Microsoft reference code) is actually the REFSYM's name. the strange thing is that REFSYM is the only type where the name is *NOT* part of the symbol length. it is however in REFSYM2, which is just REFSYM with a null terminated string instead of the length prefixed (ST) one
there are other _ST/non _ST field combinations, but they don't exhibit this behavioral difference in the length calculation.
it really feels like Microsoft somehow forgot to bump the length field when they introduced REFSYM, and then fixed it up in REFSYM2.

therefore, we need to introduce a hack (essentially a port of `cbForSym`), to add REFSYM's trailing string to the symbol data size